restructure project and rename files
