+++ lib/variables.py
@@ -1,6 +1,6 @@
-
 class Variables:
 	pass
+
 
 Variables.tileSize = 32 # 16 or 32
 Variables.SLEEP_TIME = 25
@@ -13,4 +13,3 @@ Variables.S_PLAYER = "S"
 Variables.S_PLAYER_AND_GOLD = "A"
 Variables.S_PLAYER_AND_MISSING_GOLD = 'H'
 Variables.S_MISSING_GOLD = 'G'
-
